move state objects up a level
