+++ lib/agent0/agent0/hyperdrive/agents/hyperdrive_account.py
@@ -11,8 +11,7 @@ from fixedpointmath import FixedPoint
 from agent0.base import MarketType, Trade
 from agent0.base.agents import EthAgent
 from agent0.base.policies import BasePolicy
-from agent0.hyperdrive.interface import HyperdriveReadInterface
-from agent0.hyperdrive.state import HyperdriveMarketAction, HyperdriveWallet
+from agent0.hyperdrive import HyperdriveMarketAction, HyperdriveReadInterface, HyperdriveWallet
 
 Policy = TypeVar("Policy", bound=BasePolicy)
