+++ lib/agent0/agent0/hyperdrive/exec/run_agents.py
@@ -18,8 +18,7 @@ from hexbytes import HexBytes
 
 from agent0.base import Quantity, TokenType
 from agent0.base.config import DEFAULT_USERNAME
-from agent0.hyperdrive.interface import HyperdriveReadWriteInterface
-from agent0.hyperdrive.state import HyperdriveWallet, Long, Short
+from agent0.hyperdrive import HyperdriveReadWriteInterface, HyperdriveWallet, Long, Short
 
 from .create_and_fund_user_account import create_and_fund_user_account
 from .fund_agents import async_fund_agents
