+++ lib/agent0/agent0/hyperdrive/exec/trade_loop.py
@@ -9,10 +9,9 @@ from datetime import datetime
 from web3 import Web3
 from web3.types import RPCEndpoint
 
+from agent0.hyperdrive import HyperdriveReadInterface, HyperdriveReadWriteInterface, TradeResult, TradeStatus
 from agent0.hyperdrive.agents import HyperdriveAgent
 from agent0.hyperdrive.crash_report import get_anvil_state_dump, log_hyperdrive_crash_report
-from agent0.hyperdrive.interface import HyperdriveReadInterface, HyperdriveReadWriteInterface
-from agent0.hyperdrive.state import TradeResult, TradeStatus
 from agent0.test_utils import assert_never
 
 from .execute_agent_trades import async_execute_agent_trades
