+++ lib/agent0/agent0/hyperdrive/interactive/interactive_hyperdrive.py
@@ -19,16 +19,29 @@ from chainsync.dashboard.usernames import build_user_mapping
 from chainsync.db.base import add_addr_to_username, get_addr_to_username, get_username_to_user, initialize_session
 from chainsync.db.hyperdrive import get_checkpoint_info
 from chainsync.db.hyperdrive import get_current_wallet as chainsync_get_current_wallet
-from chainsync.db.hyperdrive import (get_latest_block_number_from_analysis_table, get_pool_analysis, get_pool_config,
-                                     get_pool_info, get_ticker, get_total_wallet_pnl_over_time, get_wallet_deltas,
-                                     get_wallet_pnl)
+from chainsync.db.hyperdrive import (
+    get_latest_block_number_from_analysis_table,
+    get_pool_analysis,
+    get_pool_config,
+    get_pool_info,
+    get_ticker,
+    get_total_wallet_pnl_over_time,
+    get_wallet_deltas,
+    get_wallet_pnl,
+)
 from chainsync.exec import acquire_data, data_analysis
 from eth_account.account import Account
 from eth_typing import BlockNumber, ChecksumAddress
 from ethpy import EthConfig
 from ethpy.base import set_anvil_account_balance, smart_contract_transact
-from ethpy.hyperdrive import (BASE_TOKEN_SYMBOL, AssetIdPrefix, DeployedHyperdrivePool, ReceiptBreakdown,
-                              deploy_hyperdrive_from_factory, encode_asset_id)
+from ethpy.hyperdrive import (
+    BASE_TOKEN_SYMBOL,
+    AssetIdPrefix,
+    DeployedHyperdrivePool,
+    ReceiptBreakdown,
+    deploy_hyperdrive_from_factory,
+    encode_asset_id,
+)
 from fixedpointmath import FixedPoint
 from hypertypes import FactoryConfig, Fees, PoolDeployConfig
 from numpy.random._generator import Generator
@@ -38,17 +51,24 @@ from web3.constants import ADDRESS_ZERO
 from web3.exceptions import TimeExhausted
 
 from agent0.base.make_key import make_private_key
+from agent0.hyperdrive import HyperdriveActionType, HyperdriveReadWriteInterface, TradeResult, TradeStatus
 from agent0.hyperdrive.agents import HyperdriveAgent
 from agent0.hyperdrive.crash_report import get_anvil_state_dump, log_hyperdrive_crash_report
 from agent0.hyperdrive.exec import async_execute_agent_trades, build_wallet_positions_from_data, set_max_approval
-from agent0.hyperdrive.interface import HyperdriveReadWriteInterface
 from agent0.hyperdrive.policies import HyperdrivePolicy
-from agent0.hyperdrive.state import HyperdriveActionType, TradeResult, TradeStatus
 from agent0.test_utils import assert_never
 
 from .chain import Chain
-from .event_types import (AddLiquidity, CloseLong, CloseShort, CreateCheckpoint, OpenLong, OpenShort,
-                          RedeemWithdrawalShares, RemoveLiquidity)
+from .event_types import (
+    AddLiquidity,
+    CloseLong,
+    CloseShort,
+    CreateCheckpoint,
+    OpenLong,
+    OpenShort,
+    RedeemWithdrawalShares,
+    RemoveLiquidity,
+)
 from .interactive_hyperdrive_agent import InteractiveHyperdriveAgent
 from .interactive_hyperdrive_policy import InteractiveHyperdrivePolicy
