+++ lib/agent0/agent0/hyperdrive/interactive/interactive_hyperdrive_agent.py
@@ -11,8 +11,8 @@ if TYPE_CHECKING:
 
     from eth_typing import ChecksumAddress
 
+    from agent0.hyperdrive import HyperdriveWallet
     from agent0.hyperdrive.policies import HyperdrivePolicy
-    from agent0.hyperdrive.state import HyperdriveWallet
 
     from .event_types import (
         AddLiquidity,
