+++ lib/agent0/agent0/hyperdrive/interactive/interactive_hyperdrive_policy.py
@@ -9,9 +9,8 @@ from typing import Type
 from fixedpointmath import FixedPoint
 
 from agent0.base import MarketType, Trade
-from agent0.hyperdrive.interface import HyperdriveReadInterface
+from agent0.hyperdrive import HyperdriveActionType, HyperdriveMarketAction, HyperdriveReadInterface, HyperdriveWallet
 from agent0.hyperdrive.policies import HyperdrivePolicy
-from agent0.hyperdrive.state import HyperdriveActionType, HyperdriveMarketAction, HyperdriveWallet
 
 
 class InteractiveHyperdrivePolicy(HyperdrivePolicy):
