+++ lib/agent0/agent0/hyperdrive/interactive/interactive_hyperdrive_test.py
@@ -9,9 +9,9 @@ from ethpy.hyperdrive import BASE_TOKEN_SYMBOL, AssetIdPrefix, encode_asset_id
 from fixedpointmath import FixedPoint
 from pandas import Series
 
+from agent0.hyperdrive import HyperdriveWallet
 from agent0.hyperdrive.interactive.chain import Chain
 from agent0.hyperdrive.policies import Zoo
-from agent0.hyperdrive.state import HyperdriveWallet
 
 from .chain import LocalChain
 from .interactive_hyperdrive import InteractiveHyperdrive
