+++ lib/agent0/agent0/hyperdrive/interface/read_interface.py
@@ -23,7 +23,7 @@ from hypertypes import ERC20MintableContract, HyperdriveFactoryContract, IERC462
 from web3.types import BlockData, BlockIdentifier, Timestamp
 
 from agent0.base import MarketType, Trade
-from agent0.hyperdrive.state import HyperdriveActionType, HyperdriveMarketAction
+from agent0.hyperdrive import HyperdriveActionType, HyperdriveMarketAction
 
 from ._block_getters import _get_block, _get_block_number, _get_block_time
 from ._contract_calls import (
