+++ lib/agent0/agent0/hyperdrive/policies/arbitrage.py
@@ -8,13 +8,12 @@ from typing import TYPE_CHECKING
 from fixedpointmath import FixedPoint
 
 from agent0.base import Trade
-from agent0.hyperdrive.state import HyperdriveMarketAction
+from agent0.hyperdrive import HyperdriveMarketAction
 
 from .hyperdrive_policy import HyperdrivePolicy
 
 if TYPE_CHECKING:
-    from agent0.hyperdrive.interface import HyperdriveReadInterface
-    from agent0.hyperdrive.state import HyperdriveWallet
+    from agent0.hyperdrive import HyperdriveReadInterface, HyperdriveWallet
 
 
 class Arbitrage(HyperdrivePolicy):
