+++ lib/agent0/agent0/hyperdrive/policies/hyperdrive_policy.py
@@ -2,8 +2,7 @@
 
 from agent0.base import Trade
 from agent0.base.policies import BasePolicy
-from agent0.hyperdrive.interface import HyperdriveReadInterface
-from agent0.hyperdrive.state import HyperdriveMarketAction, HyperdriveWallet
+from agent0.hyperdrive import HyperdriveMarketAction, HyperdriveReadInterface, HyperdriveWallet
 
 
 class HyperdrivePolicy(BasePolicy[HyperdriveReadInterface, HyperdriveWallet]):
