+++ lib/agent0/agent0/hyperdrive/policies/smart_long.py
@@ -8,13 +8,12 @@ from typing import TYPE_CHECKING
 from fixedpointmath import FixedPoint, FixedPointMath
 
 from agent0.base import WEI, Trade
-from agent0.hyperdrive.state import HyperdriveMarketAction
+from agent0.hyperdrive import HyperdriveMarketAction
 
 from .hyperdrive_policy import HyperdrivePolicy
 
 if TYPE_CHECKING:
-    from agent0.hyperdrive.interface import HyperdriveReadInterface
-    from agent0.hyperdrive.state import HyperdriveWallet
+    from agent0.hyperdrive import HyperdriveReadInterface, HyperdriveWallet
 # pylint: disable=too-few-public-methods
