+++ lib/agent0/bin/fuzz_bot_invariant_checks.py
@@ -31,8 +31,8 @@ from web3.exceptions import BlockNotFound
 from web3.types import BlockData
 
 from agent0.base.config import EnvironmentConfig
+from agent0.hyperdrive import HyperdriveReadInterface
 from agent0.hyperdrive.crash_report import build_crash_trade_result, get_anvil_state_dump, log_hyperdrive_crash_report
-from agent0.hyperdrive.interface import HyperdriveReadInterface
 from agent0.interactive_fuzz.helpers import FuzzAssertionException, fp_isclose
