+++ lib/agent0/bin/fuzz_bots.py
@@ -12,8 +12,8 @@ from web3.types import RPCEndpoint
 
 from agent0 import initialize_accounts
 from agent0.base.config import AgentConfig, EnvironmentConfig
+from agent0.hyperdrive import HyperdriveReadInterface
 from agent0.hyperdrive.exec import setup_and_run_agent_loop
-from agent0.hyperdrive.interface import HyperdriveReadInterface
 from agent0.hyperdrive.policies import Zoo
 
 STOP_CHAIN_ON_CRASH = False
