+++ lib/agent0/examples/custom_agent.py
@@ -12,13 +12,12 @@ from fixedpointmath import FixedPoint
 from agent0 import initialize_accounts
 from agent0.base import Trade
 from agent0.base.config import AgentConfig, EnvironmentConfig
+from agent0.hyperdrive import HyperdriveMarketAction
 from agent0.hyperdrive.exec import setup_and_run_agent_loop
 from agent0.hyperdrive.policies import HyperdrivePolicy
-from agent0.hyperdrive.state import HyperdriveMarketAction
 
 if TYPE_CHECKING:
-    from agent0.hyperdrive.interface import HyperdriveReadInterface
-    from agent0.hyperdrive.state import HyperdriveWallet
+    from agent0.hyperdrive import HyperdriveReadInterface, HyperdriveWallet
 
 # %%
 # Define the unique agent env filename to use for this script
