+++ tests/bot_to_db_test.py
@@ -31,8 +31,8 @@ from web3 import HTTPProvider
 
 from agent0 import build_account_key_config_from_agent_config
 from agent0.base.config import AgentConfig, EnvironmentConfig
+from agent0.hyperdrive import HyperdriveReadInterface
 from agent0.hyperdrive.exec import setup_and_run_agent_loop
-from agent0.hyperdrive.interface import HyperdriveReadInterface
 from agent0.test_utils import CycleTradesPolicy
